gui/requirements-gui.txt
# GUI Requirements for RL Racing Control Center
PyQt6>=6.4.0
ruamel.yaml>=0.17.0
matplotlib>=3.5.0
pygame>=2.1.0  # For joystick support
